+++ modules/apps/data-engine/data-engine-rest-impl/src/test/java/com/liferay/data/engine/rest/internal/rule/function/v1_0/util/BaseDataRuleFunctionTest.java
@@ -31,7 +31,7 @@ import org.junit.Before;
 /**
  * @author Marcelo Mello
  */
-public abstract class BaseDataDefinitionRulesTestCase {
+public class BaseDataRuleFunctionTest {
 
 	@AfterClass
 	public static void tearDownClass() {
